Pass plugin search root, out location to new/guess
The initial design of this interface was a bit unclear, as it implicitly
copied the plugin's absolute path to the plugin name before loading it.
Now, the plugin location is saved separately from the name in the Plugin
object, and it is more clear who "owns" strings passed into these
functions.