C89-compatible variable declaration
Yeah, this is kind of a pain in the ass, but it will make the code
easier to compile on new platforms. C99 is widely supported on unix, but
not so much on Windows. Plus, this is the technically correct way to do
it in ANSI-C.

Because the github client for Windows doesn't support staged commits,
there are also a number of other compatibility fixes for Windows bundled
in here. Deal with it.