Always search current directory first (fixes #52)
MrsWatson needs to act less like a sequencer and more like a
command-line debugging tool. A very common use-case is to run MrsWatson
from the plugin's build directory, so rather than having a bunch of
search paths to likely system locations, or making the user specify the
desired path in --root-path, we search the current directory first.