32-bit filenames on Win32, fix memory errors
Unlike in the unix variant, found filenames were allocated on the stack
and not as dangling pointers. So the runtime was likely to overwrite
them after this function returned, which meant that listing plugins was
basically broken.

This also means that there is a memory leak here; the list itself is
freed but not the actual strings, as they are not CharStrings. The
solution here is probably just to refactor and use CharStrings, which we
can manage ourselves.