Better workaround for VST hardcoded-stereo bug
Instead of forcing all SampleBuffer objects to have 2 channels, we
represent them internally with the correct channel count but always
allocate 2 chanels. This is to prevent VST plugins from segfaulting if
they try to read the second channel from a mono source, despite the fact
that MrsWatson sets the channel count via the effSetSpeakerArrangement
dispatcher call.