Fix signal handling in Windows
I guess Windows actually does support the signal() call after all.
However it doesn't support all of the same signals that unix does, so we
have to check that each one is defined before adding a handler. A bit
verbose I guess, but it will make the code more portable.