Added experimental support for plugin editor
This is actually a modified version of the very old "editor" branch,
which due to all the changes in master could not be easily rebased. As
such, the functional changes have been ported and force-pushed to
master. Yay!

However, it is untested on Linux, not implemented on Windows, and will
crash on exit under Mac. For this reason, it is only enabled in DEBUG
builds.